download-service/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==24.10.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==14.1
wsproto==1.2.0
yt-dlp==2025.05.22
python-dotenv==1.0.1